requirements.txt
ultralytics
fastapi
uvicorn
pydantic
pyyaml
websockets
python-dotenv
python-multipart